add simulate

+++ koku/masu/processor/expired_data_remover.py
@@ -157,10 +157,11 @@ class ExpiredDataRemover:
                 removed_data = self._cleaner.purge_expired_line_item(expired_date=expiration_date, simulate=simulate)
             else:
                 removed_data = self._cleaner.purge_expired_report_data(expired_date=expiration_date, simulate=simulate)
-                if not line_items_only:
+                if not line_items_only and not simulate:
                     expiration_date = self._calculate_expiration_date()
                     CostUsageReportManifest.objects.filter(
                         provider__type=self._provider, billing_period_start_datetime__lt=expiration_date
                     ).delete()
+                    LOG.info("Deleted expired CostUsageReportManifests of provider type: %s" % (self._provider))
 
         return removed_data
